+++ CHANGES.rst
@@ -5,6 +5,7 @@ UNRELEASED
 ----------
 - Add support for camera device 3 (no prepare methods).
 - Make auto drift tracker during synchronized acquisition optional (default off).
+- Improve error recovery and notification during acquisition errors.
 
 0.20.0 (2011-11-12)
 -------------------